+++ pkg/framework/plugins/datalayer/pollingsource/plugin_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package pollingmodels
+package pollingsource
 
 import (
 	"context"
